- Added loud linker error when digit mismatch between headers and Lib/DLL arises. - Removed comparison functions of e_float with INT32.

+++ sandbox/e_float/libs/e_float/src/e_float/efx/e_float_efx.cpp
@@ -81,7 +81,7 @@ namespace
       }
 
       // Ensure that the value is normalized and adjust the exponent.
-      u |= static_cast<UINT64>(static_cast<UINT64>(1uLL) << (std::numeric_limits<native_float_type>::digits - 1));
+      u |= static_cast<UINT64>(1uLL << (std::numeric_limits<native_float_type>::digits - 1));
       e -= 1;
     }
   };
